Hermes - support repeater lists from Tapestry, add Ajax queries for HO

+++ application/FSPOC-WebProject/Controllers/Tapestry/RunController.cs
@@ -79,7 +79,7 @@ namespace FSS.Omnius.Controllers.Tapestry
                             foreach (var entitronRow in entitronRowList)
                             {
                                 var newRow = dataSource.NewRow();
-                                newRow["hiddenId"] = (int)entitronRow["id"];
+                                newRow["hiddenId"] = entitronRow["id"];
                                 foreach (var columnName in entitronRow.getColumnNames())
                                 {
                                     if (!columnsCreated)
